Use BIGINT for all handles, that makes databases easier.
git-svn-id: https://darwin.bournemouth.ac.uk/pdvrieze/svn/darwin/trunk/sql@295 3bae1aa3-1c52-4c45-a669-1fb9b31e7360

+++ darwin-sql/processengine/processinstances.sql
@@ -5,7 +5,7 @@ DROP TABLE `nodedata`;
 CREATE TABLE `processinstances` (
   `pihandle` BIGINT NOT NULL AUTO_INCREMENT,
   `owner` varchar(30) NOT NULL,
-  `pmhandle` INT NOT NULL,
+  `pmhandle` BIGINT NOT NULL,
   INDEX ( `owner` ),
   PRIMARY KEY ( `pihandle` ),
   FOREIGN KEY ( `pmhandle` ) REFERENCES `processmodels` ( `pmhandle` )
